requirements.txt
wheel
pandas
python-binance
ta
